share activity tree between activity dumps and security profiles
