+++ pkg/security/security_profile/activity_tree/activity_tree_secl.go
@@ -6,56 +6,37 @@
 //go:build linux
 // +build linux
 
-package dump
+package activity_tree
 
 import (
-	"bytes"
 	"fmt"
-	"text/template"
 
+	"github.com/DataDog/datadog-agent/pkg/security/secl/compiler/eval"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/model"
 	"github.com/DataDog/datadog-agent/pkg/security/utils"
 )
 
-// Profile holds the list of rules generated from an activity dump
-type Profile struct {
+// SeclEncoding holds the list of rules generated from an activity dump
+type SeclEncoding struct {
 	Name     string
 	Selector string
-	Rules    []ProfileRule
+	Rules    []eval.Rule
 }
 
-// ProfileRule contains the data required to generate a rule
-type ProfileRule struct {
-	ID         string
-	Expression string
-}
-
-// ProfileTemplate is the template used to generate profiles
-var ProfileTemplate = `---
-name: {{ .Name }}
-selector:
-  - {{ .Selector }}
-
-rules:{{ range .Rules }}
-  - id: {{ .ID }}
-    expression: {{ .Expression }}
-{{ end }}
-`
-
-// NewProfileRule returns a new ProfileRule
-func NewProfileRule(expression string, ruleIDPrefix string) ProfileRule {
-	return ProfileRule{
+// NewSeclRule returns a new ProfileRule
+func NewSeclRule(expression string, ruleIDPrefix string) eval.Rule {
+	return eval.Rule{
 		ID:         ruleIDPrefix + "_" + utils.RandString(5),
 		Expression: expression,
 	}
 }
 
-func (ad *ActivityDump) generateDNSRule(dns *DNSNode, activityNode *ProcessActivityNode, ancestors []*ProcessActivityNode, ruleIDPrefix string) []ProfileRule {
-	var rules []ProfileRule
+func (at *ActivityTree) generateDNSRule(dns *DNSNode, activityNode *ProcessNode, ancestors []*ProcessNode, ruleIDPrefix string) []eval.Rule {
+	var rules []eval.Rule
 
 	if dns != nil {
 		for _, req := range dns.Requests {
-			rule := NewProfileRule(fmt.Sprintf(
+			rule := NewSeclRule(fmt.Sprintf(
 				"dns.question.name == \"%s\" && dns.question.type == \"%s\"",
 				req.Name,
 				model.QType(req.Type).String()),
@@ -72,22 +53,22 @@ func (ad *ActivityDump) generateDNSRule(dns *DNSNode, activityNode *ProcessActiv
 	return rules
 }
 
-func (ad *ActivityDump) generateBindRule(sock *SocketNode, activityNode *ProcessActivityNode,
-	ancestors []*ProcessActivityNode, ruleIDPrefix string) []ProfileRule {
-	var rules []ProfileRule
+func (at *ActivityTree) generateBindRule(sock *SocketNode, activityNode *ProcessNode,
+	ancestors []*ProcessNode, ruleIDPrefix string) []eval.Rule {
+	var rules []eval.Rule
 
 	if sock != nil {
-		var socketRules []ProfileRule
+		var socketRules []eval.Rule
 		if len(sock.Bind) > 0 {
 			for _, bindNode := range sock.Bind {
-				socketRules = append(socketRules, NewProfileRule(fmt.Sprintf(
+				socketRules = append(socketRules, NewSeclRule(fmt.Sprintf(
 					"bind.addr.family == %s && bind.addr.ip in %s/32 && bind.addr.port == %d",
 					sock.Family, bindNode.IP, bindNode.Port),
 					ruleIDPrefix,
 				))
 			}
 		} else {
-			socketRules = []ProfileRule{NewProfileRule(fmt.Sprintf("bind.addr.family == %s", sock.Family),
+			socketRules = []eval.Rule{NewSeclRule(fmt.Sprintf("bind.addr.family == %s", sock.Family),
 				ruleIDPrefix,
 			)}
 		}
@@ -106,15 +87,15 @@ func (ad *ActivityDump) generateBindRule(sock *SocketNode, activityNode *Process
 	return rules
 }
 
-func (ad *ActivityDump) generateFIMRules(file *FileActivityNode, activityNode *ProcessActivityNode, ancestors []*ProcessActivityNode, ruleIDPrefix string) []ProfileRule {
-	var rules []ProfileRule
+func (at *ActivityTree) generateFIMRules(file *FileNode, activityNode *ProcessNode, ancestors []*ProcessNode, ruleIDPrefix string) []eval.Rule {
+	var rules []eval.Rule
 
 	if file.File == nil {
 		return rules
 	}
 
 	if file.Open != nil {
-		rule := NewProfileRule(fmt.Sprintf(
+		rule := NewSeclRule(fmt.Sprintf(
 			"open.file.path == \"%s\" && open.file.in_upper_layer == %v && open.file.uid == %d && open.file.gid == %d",
 			file.File.PathnameStr,
 			file.File.InUpperLayer,
@@ -130,18 +111,18 @@ func (ad *ActivityDump) generateFIMRules(file *FileActivityNode, activityNode *P
 	}
 
 	for _, child := range file.Children {
-		childrenRules := ad.generateFIMRules(child, activityNode, ancestors, ruleIDPrefix)
+		childrenRules := at.generateFIMRules(child, activityNode, ancestors, ruleIDPrefix)
 		rules = append(rules, childrenRules...)
 	}
 
 	return rules
 }
 
-func (ad *ActivityDump) generateRules(node *ProcessActivityNode, ancestors []*ProcessActivityNode, ruleIDPrefix string) []ProfileRule {
-	var rules []ProfileRule
+func (at *ActivityTree) generateRules(node *ProcessNode, ancestors []*ProcessNode, ruleIDPrefix string) []eval.Rule {
+	var rules []eval.Rule
 
 	// add exec rule
-	rule := NewProfileRule(fmt.Sprintf(
+	rule := NewSeclRule(fmt.Sprintf(
 		"exec.file.path == \"%s\" && process.uid == %d && process.gid == %d && process.cap_effective == %d && process.cap_permitted == %d",
 		node.Process.FileEvent.PathnameStr,
 		node.Process.UID,
@@ -157,26 +138,26 @@ func (ad *ActivityDump) generateRules(node *ProcessActivityNode, ancestors []*Pr
 
 	// add FIM rules
 	for _, file := range node.Files {
-		fimRules := ad.generateFIMRules(file, node, ancestors, ruleIDPrefix)
+		fimRules := at.generateFIMRules(file, node, ancestors, ruleIDPrefix)
 		rules = append(rules, fimRules...)
 	}
 
 	// add DNS rules
 	for _, dns := range node.DNSNames {
-		dnsRules := ad.generateDNSRule(dns, node, ancestors, ruleIDPrefix)
+		dnsRules := at.generateDNSRule(dns, node, ancestors, ruleIDPrefix)
 		rules = append(rules, dnsRules...)
 	}
 
 	// add Bind rules
 	for _, sock := range node.Sockets {
-		bindRules := ad.generateBindRule(sock, node, ancestors, ruleIDPrefix)
+		bindRules := at.generateBindRule(sock, node, ancestors, ruleIDPrefix)
 		rules = append(rules, bindRules...)
 	}
 
 	// add children rules recursively
-	newAncestors := append([]*ProcessActivityNode{node}, ancestors...)
+	newAncestors := append([]*ProcessNode{node}, ancestors...)
 	for _, child := range node.Children {
-		childrenRules := ad.generateRules(child, newAncestors, ruleIDPrefix)
+		childrenRules := at.generateRules(child, newAncestors, ruleIDPrefix)
 		rules = append(rules, childrenRules...)
 	}
 
@@ -184,34 +165,17 @@ func (ad *ActivityDump) generateRules(node *ProcessActivityNode, ancestors []*Pr
 }
 
 // GenerateProfileData generates a Profile from the activity dump
-func (ad *ActivityDump) GenerateProfileData() Profile {
-	ad.Lock()
-	defer ad.Unlock()
-
-	p := Profile{
-		Name: "profile_" + utils.RandString(5),
-	}
-
-	// generate selector
-	if len(ad.Metadata.Comm) > 0 {
-		p.Selector = fmt.Sprintf("process.comm = \"%s\"", ad.Metadata.Comm)
+func (at *ActivityTree) GenerateProfileData(selector string) SeclEncoding {
+	p := SeclEncoding{
+		Name:     "profile_" + utils.RandString(5),
+		Selector: selector,
 	}
 
 	// Add rules
-	for _, node := range ad.ProcessActivityTree {
-		rules := ad.generateRules(node, nil, p.Name)
+	for _, node := range at.ProcessNodes {
+		rules := at.generateRules(node, nil, p.Name)
 		p.Rules = append(p.Rules, rules...)
 	}
 
 	return p
-}
-
-// EncodeSecL encodes an activity dump in the SecL format
-func (ad *ActivityDump) EncodeSecL() (*bytes.Buffer, error) {
-	t := template.Must(template.New("tmpl").Parse(ProfileTemplate))
-	raw := bytes.NewBuffer(nil)
-	if err := t.Execute(raw, ad.GenerateProfileData()); err != nil {
-		return nil, fmt.Errorf("couldn't generate profile: %w", err)
-	}
-	return raw, nil
 }
