+++ pkg/security/security_profile/activity_tree/activity_tree_test.go
@@ -1,0 +1,68 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build linux
+// +build linux
+
+package activity_tree
+
+import (
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	"github.com/DataDog/datadog-agent/pkg/security/secl/model"
+)
+
+func TestInsertFileEvent(t *testing.T) {
+	pan := ProcessNode{
+		Files: make(map[string]*FileNode),
+	}
+	pan.Process.FileEvent.PathnameStr = "/test/pan"
+	stats := NewActivityTreeNodeStats()
+
+	pathToInserts := []string{
+		"/tmp/foo",
+		"/tmp/bar",
+		"/test/a/b/c/d/e/",
+		"/hello",
+		"/tmp/bar/test",
+	}
+	expectedDebugOuput := strings.TrimSpace(`
+- process: /test/pan
+  files:
+    - hello
+    - test
+        - a
+            - b
+                - c
+                    - d
+                        - e
+    - tmp
+        - bar
+            - test
+        - foo
+`)
+
+	for _, path := range pathToInserts {
+		event := &model.Event{
+			Open: model.OpenEvent{
+				File: model.FileEvent{
+					IsPathnameStrResolved: true,
+					PathnameStr:           path,
+				},
+			},
+			FieldHandlers: &model.DefaultFieldHandlers{},
+		}
+		pan.InsertFileEvent(&event.Open.File, event, Unknown, stats, false, false)
+	}
+
+	var builder strings.Builder
+	pan.debug(&builder, "")
+	debugOutput := strings.TrimSpace(builder.String())
+
+	assert.Equal(t, expectedDebugOuput, debugOutput)
+}
