+++ pkg/security/security_profile/dump/activity_dump_profile_proto_enc_v1.go
@@ -10,6 +10,8 @@ package dump
 
 import (
 	proto "github.com/DataDog/agent-payload/v5/cws/dumpsv1"
+
+	"github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree"
 )
 
 // ActivityDumpToSecurityProfileProto serializes an Activity Dump to a Security Profile protobuf representation
@@ -22,15 +24,11 @@ func ActivityDumpToSecurityProfileProto(input *ActivityDump) *proto.SecurityProf
 		Status:   3, // AnomalyDetection + AutoSuppression
 		Version:  "1",
 		Metadata: adMetadataToProto(&input.Metadata),
-		Syscalls: input.computeSyscallsList(),
+		Syscalls: input.ActivityTree.ComputeSyscallsList(),
 		Tags:     make([]string, len(input.Tags)),
-		Tree:     make([]*proto.ProcessActivityNode, 0, len(input.ProcessActivityTree)),
+		Tree:     activity_tree.ActivityTreeToProto(input.ActivityTree),
 	}
 	copy(output.Tags, input.Tags)
-
-	for _, tree := range input.ProcessActivityTree {
-		output.Tree = append(output.Tree, processActivityNodeToProto(tree))
-	}
 
 	return &output
 }
