+++ pkg/security/security_profile/dump/manager_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/DataDog/datadog-go/v5/statsd"
 
 	"github.com/DataDog/datadog-agent/pkg/security/config"
+	"github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree"
 )
 
 func compareListOfDumps(t *testing.T, out, expectedOut []*ActivityDump) {
@@ -221,11 +222,15 @@ func TestActivityDumpManager_getOverweightDumps(t *testing.T) {
 			"one_dump/one_overweight_dump",
 			fields{
 				activeDumps: []*ActivityDump{
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "1"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "1"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+					}},
 				},
 			},
 			[]*ActivityDump{
-				{Metadata: Metadata{Name: "1"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
+				{Metadata: Metadata{Name: "1"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+				}},
 			},
 			[]*ActivityDump{},
 		},
@@ -265,19 +270,39 @@ func TestActivityDumpManager_getOverweightDumps(t *testing.T) {
 			"5_dumps/5_overweight_dumps",
 			fields{
 				activeDumps: []*ActivityDump{
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "1"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "2"}, nodeStats: ActivityDumpNodeStats{processNodes: 3}},
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "3"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "4"}, nodeStats: ActivityDumpNodeStats{processNodes: 3}},
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "5"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "1"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+					}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "2"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 3},
+					}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "3"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+					}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "4"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 3},
+					}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "5"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+					}},
 				},
 			},
 			[]*ActivityDump{
-				{Metadata: Metadata{Name: "1"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
-				{Metadata: Metadata{Name: "2"}, nodeStats: ActivityDumpNodeStats{processNodes: 3}},
-				{Metadata: Metadata{Name: "3"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
-				{Metadata: Metadata{Name: "4"}, nodeStats: ActivityDumpNodeStats{processNodes: 3}},
-				{Metadata: Metadata{Name: "5"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
+				{Metadata: Metadata{Name: "1"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+				}},
+				{Metadata: Metadata{Name: "2"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 3},
+				}},
+				{Metadata: Metadata{Name: "3"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+				}},
+				{Metadata: Metadata{Name: "4"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 3},
+				}},
+				{Metadata: Metadata{Name: "5"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+				}},
 			},
 			[]*ActivityDump{},
 		},
@@ -286,15 +311,23 @@ func TestActivityDumpManager_getOverweightDumps(t *testing.T) {
 			fields{
 				activeDumps: []*ActivityDump{
 					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "1"}},
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "2"}, nodeStats: ActivityDumpNodeStats{processNodes: 3}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "2"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 3},
+					}},
 					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "3"}},
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "4"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "4"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+					}},
 					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "5"}},
 				},
 			},
 			[]*ActivityDump{
-				{Metadata: Metadata{Name: "2"}, nodeStats: ActivityDumpNodeStats{processNodes: 3}},
-				{Metadata: Metadata{Name: "4"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
+				{Metadata: Metadata{Name: "2"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 3},
+				}},
+				{Metadata: Metadata{Name: "4"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+				}},
 			},
 			[]*ActivityDump{
 				{Metadata: Metadata{Name: "1"}},
@@ -306,16 +339,24 @@ func TestActivityDumpManager_getOverweightDumps(t *testing.T) {
 			"5_dumps/2_expired_dumps_at_the_start",
 			fields{
 				activeDumps: []*ActivityDump{
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "1"}, nodeStats: ActivityDumpNodeStats{processNodes: 3}},
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "2"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "1"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 3},
+					}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "2"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+					}},
 					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "3"}},
 					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "4"}},
 					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "5"}},
 				},
 			},
 			[]*ActivityDump{
-				{Metadata: Metadata{Name: "1"}, nodeStats: ActivityDumpNodeStats{processNodes: 3}},
-				{Metadata: Metadata{Name: "2"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
+				{Metadata: Metadata{Name: "1"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 3},
+				}},
+				{Metadata: Metadata{Name: "2"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+				}},
 			},
 			[]*ActivityDump{
 				{Metadata: Metadata{Name: "3", End: time.Now().Add(time.Minute)}},
@@ -330,13 +371,21 @@ func TestActivityDumpManager_getOverweightDumps(t *testing.T) {
 					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "1"}},
 					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "2"}},
 					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "3"}},
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "4"}, nodeStats: ActivityDumpNodeStats{processNodes: 3}},
-					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "5"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "4"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 3},
+					}},
+					{Mutex: &sync.Mutex{}, Metadata: Metadata{Name: "5"}, ActivityTree: &activity_tree.ActivityTree{
+						Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+					}},
 				},
 			},
 			[]*ActivityDump{
-				{Metadata: Metadata{Name: "4"}, nodeStats: ActivityDumpNodeStats{processNodes: 3}},
-				{Metadata: Metadata{Name: "5"}, nodeStats: ActivityDumpNodeStats{processNodes: 2}},
+				{Metadata: Metadata{Name: "4"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 3},
+				}},
+				{Metadata: Metadata{Name: "5"}, ActivityTree: &activity_tree.ActivityTree{
+					Stats: &activity_tree.ActivityTreeStats{ProcessNodes: 2},
+				}},
 			},
 			[]*ActivityDump{
 				{Metadata: Metadata{Name: "1"}},
